Set PublishAot instead of IsTrimmable
