+++ src/Sentry/Sentry.csproj
@@ -23,7 +23,7 @@
   <Import Project="Platforms\iOS\Sentry.iOS.props" Condition="'$(TargetPlatformIdentifier)' == 'ios' Or '$(TargetPlatformIdentifier)' == 'maccatalyst'" />
   <Import Project="Platforms\Native\Sentry.Native.props" Condition="('$(TargetFramework)' == 'net7.0' Or '$(TargetFramework)' == 'net8.0') And ($([MSBuild]::IsOSPlatform('Linux')) Or $([MSBuild]::IsOSPlatform('Windows')))" />
 
-  <Target Name="IncludeBenDemystifier" BeforeTargets="BeforeBuild" Condition="'$(IsTrimmable)' != 'true'">
+  <Target Name="IncludeBenDemystifier" BeforeTargets="BeforeBuild" Condition="'$(PublishAot)' != 'true'">
     <!--
       Ben.Demystifier is compiled directly into Sentry.
       Note: It uses Microsoft.Bcl.AsyncInterfaces, which we get transitively from System.Text.Json.
